Вынесла функции сортировки игроков по уровню Golevel, сортировки по рангу, а также функцию по нахождения позиции игрока в КГФ в отдельный файл. Вывела данные в шаблон, сделала апдейт представления для отображения страницы Игроки
