+++ source/webapp/templates/partial/player_list.html
@@ -9,39 +9,36 @@
         <th scope="col">Age</th>
         <th scope="col">Club</th>
         <th scope="col">Rank</th>
-        <th scope="col">Tot/trnaments</th>
+        <th scope="col">Tot/tournaments</th>
     </tr>
     </thead>
     <tbody>
-    {% for player in players %}
+    {% for el in sorted_players %}
         <tr>
             <td>{{ forloop.counter }}</td>
             <td><a style="text-decoration: underline dotted ghostwhite; color: black;"
-                   href="{% url "webapp:player_detail" player.pk %}">{{ player.last_name }}</a></td>
-            <td>{{ player.first_name }}</td>
-            {% if player.patronymic is not None %}
-                <td>{{ player.patronymic }}</td>
+                   href="{% url "webapp:player_detail" el.player.pk %}">{{ el.player.last_name }}</a></td>
+            <td>{{ el.player.first_name }}</td>
+            {% if el.player.patronymic is not None %}
+                <td>{{ el.player.patronymic }}</td>
             {% else %}
                 <td></td>
             {% endif %}
-            {% if player.birth_date %}
-                <td>{{ player.get_age_date }}</td>
+            {% if el.player.birth_date %}
+                <td>{{ el.player.get_age_date }}</td>
             {% else %}
                 <td></td>
             {% endif %}
-            {% if player.clubs.all %}
-                {% for club in player.clubs.all %}
+            {% if el.player.clubs.all %}
+                {% for club in el.player.clubs.all %}
                     <td>{{ club.name }}</td>
                 {% endfor %}
             {% else %}
-                <td>Не состоит в клубах</td>
+                <td></td>
             {% endif %}
-            {% for element in rank %}
-                {% if player.pk == element.player %}
-                    <td>{{ element.GoLevel }}</td>
-                {% endif %}
-            {% endfor %}
-            <td>{{ player.get_total_tournaments }}</td>
+            <td>{{ el.GoLevel }}</td>
+
+            <td>{{ el.player.get_total_tournaments }}</td>
         </tr>
     {% endfor %}
     </tbody>
