+++ source/webapp/views/__init__.py
@@ -4,3 +4,4 @@ from .club_view import *
 from .calendar_views import *
 from .tournament_view import *
 from .news_views import *
+from .functions import *
